# docs/swagger.yaml
basePath: /
host: localhost:5000
info:
  contact:
    name: API Support
  description: swagger Projeto Raul
  title: Fiber Example API
  version: "1.0"
paths:
...
  /creature:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Creature
  /creatures:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Creature
  /decoy:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Decoy
  /decoys:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Decoy
...
  /deleteCreature/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Creature by its ID
      operationId: delete-creature-by-int
      parameters:
      - description: Creature ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Creature
      tags:
      - Creature
  /deleteDecoy/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Decoy by its ID
      operationId: delete-decoy-by-int
      parameters:
      - description: Decoy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Decoy
      tags:
      - Decoy
  /deleteEndemicCreature/{id}:
    delete:
      consumes:
      - application/json
      description: delete a EndemicCreature by its ID
      operationId: delete-endemicCreature-by-int
      parameters:
      - description: EndemicCreature ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a EndemicCreature
      tags:
      - EndemicCreature
  /deleteEndemicVegetation/{id}:
    delete:
      consumes:
      - application/json
      description: delete a EndemicVegetation by its ID
      operationId: delete-endemicVegetations-by-int
      parameters:
      - description: EndemicVegetation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a EndemicVegetation
      tags:
      - EndemicVegetation
...
  /endemicCreature:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new EndemicCreature
  /endemicCreatures:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of EndemicCreature
  /endemicVegetation:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new EndemicVegetation
  /endemicVegetations:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of EndemicVegetation
...
  /updateCreature/{id}:
    put:
      consumes:
      - application/json
      description: update a Creature by its ID
      operationId: get-Creature-by-int
      parameters:
      - description: Creature ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Creature
      tags:
      - Item
  /updateDecoy/{id}:
    put:
      consumes:
      - application/json
      description: update a Decoy by its ID
      operationId: get-Decoy-by-int
      parameters:
      - description: Decoy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Decoy
      tags:
      - Item
  /updateEndemicCreature/{id}:
    put:
      consumes:
      - application/json
      description: update a EndemicCreature by its ID
      operationId: get-EndemicCreature-by-int
      parameters:
      - description: EndemicCreature ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a EndemicCreature
      tags:
      - Item
  /updateEndemicVegetation/{id}:
    put:
      consumes:
      - application/json
      description: update a EndemicVegetation by its ID
      operationId: get-EndemicVegetation-by-int
      parameters:
      - description: EndemicVegetation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a EndemicVegetation
      tags:
      - Item
...
swagger: "2.0"